Fix 64-bit integer/pointer cast sites in shared compiler code
Replaces int32-to-void* casts in print calls with properly sized
casts and routes pointer-width values through intptr_t. These were
latent truncation bugs inherited from the 32-bit VM.

+++ vm64/src/any/objects/oop.cpp
@@ -497,7 +497,7 @@ oop oopClass::unwind_protect_prim(oop doBlock, oop protectBlock) {
 
     LOG_EVENT2("*3* continuing: NLR=%#lx, ID=%d",
                (unsigned long)NLRSupport::NLR_home_from_C(),
-               NLRSupport::NLR_home_ID_from_C());
+               (long)NLRSupport::NLR_home_ID_from_C());
 
   } // end scope for the resource mark
   // return non-locally
